test: decompose Harbor policy tests
